multitask_cnn_rnn/requirements.txt
torch~=1.7.1
torchvision~=0.8.2
torchaudio~=0.7.2
numpy~=1.19.3
pandas~=1.1.2
Pillow~=7.2.0
matplotlib~=3.3.2
six~=1.15.0
scikit-learn~=0.22.2.post1
tqdm~=4.48.2